# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: send-flow API Blueprint
  version: 0.1.0
  description: >
    Contract-first OpenAPI specification derived from docs/api/* only.
    This file documents the product blueprint for send-flow and may include
    endpoints that are not yet implemented in code.
servers:
  - url: /
...
security: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Bearer token used by dashboard sessions and workspace API keys depending
        on the endpoint. Endpoints that already contain `workspace_id` in the
        path remain stateless by path, and dashboard-session clients should
        provide workspace context explicitly per request when needed.
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: WorkspaceApiKey
      description: Workspace API key sent as a bearer credential, for example `Bearer sk_live_xxx`.
    providerSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: Provider-specific webhook signature header.
  parameters:
...
    WorkspaceId:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
      description: Workspace identifier.
...
    MessageId:
      name: message_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    Meta:
      type: object
      additionalProperties: false
      properties:
        request_id:
          type: string
        next_cursor:
          type: string
        total_estimate:
          type: integer
      required: [request_id]
...
    Envelope:
      type: object
      additionalProperties: false
      properties:
        data: {}
        meta:
          $ref: '#/components/schemas/Meta'
      required: [data, meta]
...
    MessageResponse:
      type: object
      additionalProperties: false
      properties:
        message_id:
          type: string
        recipient:
          type: string
        status:
          type: string
        provider:
          type: string
          nullable: true
        sent_at:
          type: string
          format: date-time
          nullable: true
        last_updated_at:
          type: string
          format: date-time
          nullable: true
        provider_message_id:
          type: string
          nullable: true
        timeline:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [message_id, status]
...
paths:
...
  /api/v1/transactional/messages/{message_id}:
    get:
      tags: [Delivery]
      operationId: getTransactionalMessage
      x-implementation-status: implemented
      summary: Get transactional message status
      description: |
        Possible error.code values:
        - `api_key.invalid`
        - `api_key.scope_denied`
        - `delivery.message_not_found`
      security:
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/MessageId'
      responses:
        '200':
          description: Message status
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageResponse'
                    required: [data]
...
  /api/v1/workspaces/{workspace_id}/messages/{message_id}:
    get:
      tags: [Delivery]
      operationId: getMessage
      x-implementation-status: implemented
      summary: Get a message log entry
      description: |
        Possible error.code values:
        - `auth.invalid_token`
        - `delivery.logs_read_denied`
        - `delivery.message_not_found`
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - $ref: '#/components/parameters/MessageId'
      responses:
        '200':
          description: Message detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MessageResponse'
                    required: [data]
...